pyproject.toml
[project]
name = "pyladies-seoul"
version = "1.0.0"
description = "PyLadies Seoul Official Website - A Wagtail-based multilingual CMS"
authors = [
    {name = "PyLadies Seoul", email = "seoul@pyladies.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
keywords = ["django", "wagtail", "cms", "multilingual", "pyladies"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Framework :: Wagtail",
    "Framework :: Wagtail :: 6",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    # Core Django and Wagtail
    "Django>=5.1,<5.2",
    "wagtail>=6.2,<7.0",
    # Database
    "psycopg2-binary>=2.9.7",
    "dj-database-url>=2.1.0",
    # Caching and Redis
    "redis>=5.0.1",
    "django-redis>=5.4.0",
    # Environment and Configuration
    "python-dotenv>=1.0.0",
    "django-environ>=0.11.2",
    # Monitoring and Logging
    "logfire[django]>=0.1.0",
    "sentry-sdk[django]>=1.38.0",
    # Static Files and Assets
    "whitenoise>=6.6.0",
    "django-compressor>=4.4",
    "django-sass-processor>=1.4.1",
    "libsass>=0.22.0",
    # Image Processing
    "Pillow>=10.1.0",
    "willow[heif]>=1.8.0",
    # Internationalization
    "django-modeltranslation>=0.18.11",
    # Security
    "django-csp>=3.7",
    "django-permissions-policy>=4.17.0",
    # Performance and Optimization
    "django-cachalot>=2.6.1",
    "django-debug-toolbar>=4.2.0",
    # Email
    "django-anymail>=10.2",
    # Forms and Validation
    "django-crispy-forms>=2.1",
    "crispy-bootstrap5>=0.7",
    # Utils
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    # Markdown Support
    "wagtail-markdown>=0.11.0",
    "markdown>=3.5.0",
    "pymdown-extensions>=10.3.0",
    "django-extensions>=4.1",
    "ipython>=9.4.0",
]

[project.optional-dependencies]
development = [
    # Development Tools
    "ruff>=0.1.6",
    "black>=23.11.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
    
    # Testing
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.4.0",
    "factory-boy>=3.3.0",
    "coverage>=7.3.2",
    "model-bakery>=1.17.0",
    
    # Development Server
    "django-extensions>=3.2.3",
    "Werkzeug>=3.0.1",
    
    # Documentation
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=1.3.0",
    
    # Type Checking
    "mypy>=1.7.1",
    "django-stubs>=4.2.6",
    "types-requests>=2.31.0",
    "types-redis>=4.6.0",
    
    # Browser Testing
    "playwright>=1.40.0",
    "selenium>=4.15.2",
]

production = [
    # Production Server
    "gunicorn>=21.2.0",
    "gevent>=23.9.1",
    
    # Production Monitoring
    "newrelic>=9.2.0",
    
    # Security
    "django-secure>=1.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.6",
    "black>=23.11.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.4.0",
    "factory-boy>=3.3.0",
    "coverage>=7.3.2",
    "model-bakery>=1.17.0",
    "django-extensions>=3.2.3",
    "Werkzeug>=3.0.1",
    "mypy>=1.7.1",
    "django-stubs>=4.2.6",
    "types-requests>=2.31.0",
    "types-redis>=4.6.0",
    "playwright>=1.40.0",
    "selenium>=4.15.2",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"*/migrations/*" = ["E501"]
"*/settings/*" = ["F405", "F403"]
"*/tests/*" = ["B011"]

[tool.ruff.isort]
known-django = ["django"]
known-wagtail = ["wagtail"]
known-first-party = ["apps", "config"]
section-order = ["future", "standard-library", "third-party", "django", "wagtail", "first-party", "local-folder"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = ["django"]
known_wagtail = ["wagtail"]
known_first_party = ["apps", "config"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "WAGTAIL", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
testpaths = ["tests", "apps"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--cov=apps",
    "--cov=config",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-fail-under=80",
]

[tool.coverage.run]
source = ["apps", "config"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/venv/*",
    "*/env/*",
    "*/.venv/*",
    "manage.py",
    "*/settings/*",
    "*/wsgi.py",
    "*/asgi.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
